Move filter save/load code to a separate file
